Update client error handling - Don't return promise on redirect - Add support to pass custom error handler
